Added files via Upload
This project simulates the growth of a city per development of residential, industrial, and commercial zones, and the distribution of dogs.
A configuration file is read in to initialize the city region and output its initial state. Then, the simulation will
begin. Once complete, the final region state is outputted. The functionalities of this program are laid out and organized in their respective classes.
The SimSystem class will manage the initialization of the city region, mutations onto the region per time step (called with members of the class of each zone type), and print of the region per refresh rate. Details about this class can be read in its respectful wiki.
The Cell class contains functions to control mutations to individual Cells/zones within a region. It contains the specific values of each zone as well. Pointers are utilized here in order to store the values of adjacent cells. Additionally, booleans check for specific characteristics of the surrounding zones. Details about this class can be read in its respectful wiki.
Each zone type has its own class that holds the parameters that will update a zone. The classes are Residential, Industrial, Commercial, and Adoption. Each class has commands to update the functions whereas the Cell class does the initializing and calls upon the classes. All 3 of these cells share two public variables in order to access the variables that store the city's available goods and workers. Details about this class can be read in their respectful wikis.
The Pollution Class controls the spread of pollution throughout the region. It mainly coincides with the Industrial and utilizes cell adjacencies to update the region.
In the main class the methods provided by the SimSystem class to read in the configuration file, initialize the region using the information provided by the file, and print the initial region state. It then begins the time step sequence and outputs the
final region state upon completion.
Here is a Diagram that displays the classes, their respected functions, and how they all connect